+++ pages/components/common/SearchContent.module.css
@@ -1,21 +1,3 @@
-.content {
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  width: 1060px;
-  align-items: center;
-  justify-content: center;
-  margin: 40px auto 100px;
-  gap: 40px;
-
-  @media (max-width: 1124px) {
-    width: 704px;
-  }
-  @media (max-width: 767px) {
-    width: 325px;
-  }
-}
-
 .wrapper {
   display: grid;
   grid-template-columns: 1fr 1fr 1fr;
